stable function to lib!
