+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "pingdown"
-version = "1.1.1"
+version = "1.1.2"
 authors = ["Chris233 h-chris233@qq.com"]
 edition = "2021"
 readme = "README.md"
